requirements.txt

torch==2.7.1+cu126
torchvision==0.22.1+cu126
torchaudio==2.7.1+cu126
numpy==1.26.4
scipy==1.15.3

diffusers==0.32.0
transformers==4.51.3
accelerate==1.12.0
peft==0.14.0
safetensors==0.7.0
sentencepiece==0.2.0
huggingface_hub==0.36.1
hf_transfer==0.1.9

open3d==0.19.0
pytorch3d==0.7.9
trimesh==4.11.1
viser==1.0.21

opencv-python==4.13.0.92
imageio==2.37.2
imageio-ffmpeg==0.6.0
decord==0.6.0
moviepy==1.0.3
Pillow==12.1.0

scikit-image==0.25.2
scikit-learn==1.7.2
numba==0.63.1
einops==0.7.0
pandas==2.3.3
matplotlib==3.10.8

wandb==0.24.1
deepspeed==0.18.5
bitsandbytes==0.49.1
torchao==0.10.0

flash_attn==2.7.4.post1
xformers==0.0.31.post1

qwen-vl-utils==0.0.14

easydict==1.13
timm==0.6.7
tqdm==4.67.2
gradio==6.5.1
plyfile==1.1.3
omegaconf==2.3.0
av==16.1.0

